+++ src/core/util/automation/automaton.rs
@@ -801,9 +801,9 @@ impl Sorter for InPlaceMergeSorterImpl<'_> {
         }
     }
 
-    fn swap(&mut self, i: i32, j: i32) -> Result<()> {
-        let i_start = i as usize * 4;
-        let j_start = j as usize * 4;
+    fn swap(&mut self, i: usize, j: usize) -> Result<()> {
+        let i_start = i * 4;
+        let j_start = j * 4;
         self.swap_one(i_start, j_start);
         self.swap_one(i_start + 1, j_start + 1);
         self.swap_one(i_start + 2, j_start + 2);
@@ -856,9 +856,9 @@ impl Sorter for MinMaxDestSorter<'_> {
         }
     }
 
-    fn swap(&mut self, i: i32, j: i32) -> Result<()> {
-        let i_start = 3 * i as usize;
-        let j_start = 3 * j as usize;
+    fn swap(&mut self, i: usize, j: usize) -> Result<()> {
+        let i_start = 3 * i;
+        let j_start = 3 * j;
         self.swap_one(i_start, j_start);
         self.swap_one(i_start + 1, j_start + 1);
         self.swap_one(i_start + 2, j_start + 2);
@@ -909,9 +909,9 @@ impl Sorter for DestMinMaxSorter<'_> {
         Ok(0)
     }
 
-    fn swap(&mut self, i: i32, j: i32) -> Result<()> {
-        let i_start = (3 * i) as usize;
-        let j_start = (3 * j) as usize;
+    fn swap(&mut self, i: usize, j: usize) -> Result<()> {
+        let i_start = 3 * i;
+        let j_start = 3 * j;
         self.swap_one(i_start, j_start);
         self.swap_one(i_start + 1, j_start + 1);
         self.swap_one(i_start + 2, j_start + 2);
